pitch shift implemented with a ton of noise (because phases are not synchronised), I hope it will not be noticable later

+++ include/render/fourier.h
@@ -5,7 +5,7 @@
 
 #include "main.h"
 
-void fastFourierTransform(complex double *input, uint64_t n, complex double *temp);
-void inverseFastFourierTransform(complex double *input, uint64_t n, complex double *temp);
+void fastFourierTransform(complex float *input, uint64_t n, complex float *temp);
+void inverseFastFourierTransform(complex float *input, uint64_t n, complex float *temp);
 
 #endif
